requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
ipykernel>=6.25.0
notebook>=7.0.0

